# Optimization of Tpetra init (construction of graphs etc)