--- a/include/NaluCommNeighbors.hpp
+++ b/include/NaluCommNeighbors.hpp
@@ -98,9 +98,6 @@ public:
 
   ~CommNeighbors();
 
-  const std::vector<int>& send_procs() const { return m_send_procs; }
-  const std::vector<int>& recv_procs() const { return m_recv_procs; }
-
 private:
 
   //----------------------------------------