Sorry about putting in such a large code-change. If it helps I can split it into smaller pieces for easier review...

Add communication code to send/recv off-processor column indices
so that we can construct Tpetra graphs with exact row-lengths and
use the 'StaticProfile' constructor argument which avoids significant
re-allocation during fillComplete.
Also eliminated temporary graph which was being used to import
remove column indices for the owned graph.

Had to temporarily take copies of stk files CommBufferV.hpp and
CommNeighbors.[hpp,cpp], until next time stk is updated. I made some
important performance improvements in these files. For now, they
reside in include/NaluCommBufferV.hpp, include/NaluCommNeighbors.hpp and
src/NaluCommNeighbors.C.

Added ghosting modification to PeriodicManager to ensure that ghosts
of range nodes are on the procs that share the domain nodes.

Added tpetra unit-test, which required enhancing UnitTestRealm.C to
add some YAML-node input syntax adding an equation system etc.

Had to compute local graph arrays (row-pointers and col-indices) to
give to Tpetra using setAllIndices(..) to avoid having Tpetra sort
all the rows during fillComplete.

Tried using expertStaticFillComplete. Went back to
regular fillComplete, as the 'expert' one doesn't perform any
better than regular. Consulted with Mark Hoemmen, he says the
expert one accepts a pre-constructed Importer which could save time
if we can pre-construct an Importer with our provided remote-owner
information so that Importer doesn't have to communicate. He has
just added a function to construct an Importer that way, so we can
switch to that after Trilinos updates work their way through.

Reblessed 19 tests due to changed ordering of matrix graph. All diffs
were consistent with round-off as demonstrated by Robert's experiment
of re-running tests with changed stk-mesh ordering.

Performance result: some nightly tests such as uqSlidingMeshDG and
waleElemXflowMixFrac3.5m show improvements in tpetra initialization time
around 2x. The V27 problem running on 8192cores of cori (setup courtesy of
Shreyas and Jonathan Hu) showed init time decrease from ~116s to ~75s, with
overall runtime decrease from ~375s to ~345s.