@sayerhs 2018-04-13
@alanw0 It appears that unit tests depending on TIOGA and OpenFAST have not been updated, is this just fortunate or were those libraries not enabled in your Nalu build? Since all other overset tests and ABL tests changed, I would suspect that these tests will require updates of gold files also.