@sayerhs 2018-04-13
@alanw0 Yes, that is correct. You will have to install TIOGA and OpenFAST and enable them during CMake configure phase.