@alanw0 2018-04-13
@sayerhs ok. I found tioga on GitHub and cloned it and built it, but when I try to enable it in the nalu build I get errors about "By not providing "FindTIOGA.cmake" in CMAKE_MODULE_PATH this project has
  asked CMake to find a package configuration file provided by "TIOGA", but
  CMake did not find one.".
So I'll wait for your email before proceeding further with this.