@spdomin 2018-04-19
@alanw0, let's find time to walk through this code change so that we can have more eyes on the details.