@jhux2 2018-04-24
> Consulted with Mark Hoemmen, he says the
> expert one accepts a pre-constructed Importer which could save time
> if we can pre-construct an Importer with our provided remote-owner
> information so that Importer doesn't have to communicate. He has
> just added a function to construct an Importer that way, so we can
> switch to that after Trilinos updates work their way through.

@alanw0 @mhoemmen's new Import constructor is in master, see commit trilinos/Trilinos@43094b41 and the related issue trilinos/Trilinos#2198.