@alanw0 2018-04-27
Good discussion this morning. Added a commit to change "globally owned" to "shared not owned".